Test Quatro Document
I'm mostly just creating this document to see what it looks like on GitHub. I'm trying to compare Quarto to Rmd. One aspect I'm interested in is how nicely it plays with version control and GitHub. It's very easy to view the content of my Rmd documents on GitHub. I'm checking to see how easy it is to view the content of my qmd documents as well.

Another issue I need to look into is creating templates for Quarto documents.